+++ test/transform/resource/after-delombok/UtilityClass.java
@@ -1,0 +1,26 @@
+final class UtilityClass {
+	private static String someField;
+	static void someMethod() {
+		System.out.println();
+	}
+	protected static class InnerClass {
+		private String innerInnerMember;
+	}
+	@java.lang.SuppressWarnings("all")
+	@javax.annotation.Generated("lombok")
+	private UtilityClass() {
+		throw new java.lang.UnsupportedOperationException("This is a utility class and cannot be instantiated");
+	}
+}
+class UtilityInner {
+	static class InnerInner {
+		static final class InnerInnerInner {
+			static int member;
+			@java.lang.SuppressWarnings("all")
+			@javax.annotation.Generated("lombok")
+			private InnerInnerInner() {
+				throw new java.lang.UnsupportedOperationException("This is a utility class and cannot be instantiated");
+			}
+		}
+	}
+}