+++ test/transform/resource/after-delombok/UtilityClassErrors.java
@@ -1,0 +1,12 @@
+final class UtilityClassErrors1 {
+	private static String someField;
+	protected UtilityClassErrors1() {
+	}
+	static void method() {
+		class MethodLocalClass {
+		}
+	}
+}
+enum UtilityClassErrors2 {
+;
+}